hot fixes arrows foriswrapper update

+++ ForisWrapperQt/DosWrapper.cpp
@@ -9,6 +9,10 @@
 #include <vector>
 #include <string_view>
 #include <fstream>
+#include <sstream>
+
+#include <thread>
+#include <chrono>
 
 using namespace std;
 
@@ -26,7 +30,13 @@ void file_del(const string_view& filename) {
     auto file = QFile(filename.data());
     if(file.exists()){
         if(!file.remove()){
-            throw "delete file error";
+            int tryes = 0;
+            while(tryes != 100){
+                if(file.remove()) return;
+                ++tryes;
+                std::this_thread::sleep_for(std::chrono::milliseconds(50));
+            }
+            throw "internal error: delete file error";
         }
     }
 }
@@ -47,11 +57,13 @@ string read(string_view filename) {
     return result;
 }
 
+
 string get(string_view in_name, string_view in_acc_name) {
     auto result = read(in_name);
     file_del(in_acc_name);
     return result;
 }
+
 
 void send(string_view str, string_view out_name, string_view out_acc_name) {
     write(out_name, str);
@@ -64,11 +76,6 @@ DosWrapper::DosWrapper(dos_data data, QObject *parent)
        t_data{data}
 {
     t_create_conf();
-    //t_empty_check_timer.setInterval(500);
-    //connect(&t_empty_check_timer, &QTimer::timeout, this, &DosWrapper::t_check_empty);
-    //t_empty_check_timer.start();
-
-    //connect(qApp, &QCoreApplication::aboutToQuit, this, &DosWrapper::finish);
 }
 
 DosWrapper::~DosWrapper()
@@ -83,21 +90,15 @@ void DosWrapper::run()
     connect(t_process, static_cast<void(QProcess::*)(int)>(&QProcess::finished), this, &DosWrapper::finished);
     //connect(t_process, &QProcess::errorOccurred, this, &DosWrapper::finished);
 
-
     file_del(t_data.in_acc_name);
     file_del(t_data.out_acc_name);
     file_del(t_data.out_name);
     file_del(t_data.in_name);
 
     t_watcher = new QFileSystemWatcher(this);
-
-    if(!is_exist_dir((t_data.sinchro_path + "/sinchro").c_str())){
-        auto d = QDir(t_data.sinchro_path.c_str());
-        d.mkdir("/sinchro");
-    }
     bool beingWatched = t_watcher->addPath(QString::fromStdString(t_data.sinchro_path + "/sinchro"));
     if (!beingWatched ) {
-        qDebug() << "watcher error";
+        qDebug() << "internal error: watcher error";
         emit finished();
         return;
     }
@@ -118,20 +119,22 @@ void DosWrapper::run()
     t_process->setNativeArguments(str);
     t_process->start(t_data.dos_exe_path.c_str());
 
-    if( !t_process->waitForStarted(5000)) {
+    if( !t_process->waitForStarted(3000)) {
         delete t_watcher;
         t_watcher = nullptr;
         emit finished();
     }
-    cout << "ForisWrapper 0.1 (c)" << endl;
+    cout << "ForisWrapper 0.19 (c)" << endl;
 }
 
 void DosWrapper::finish()
 {
-    t_process->terminate();
-    t_process->waitForFinished(1000);
-    if(t_process->state() == QProcess::Running)t_process->kill();
-    t_process->waitForFinished(100);
+    if(t_process && (t_process->state() == QProcess::Running)){
+        t_process->terminate();
+        t_process->waitForFinished(1000);
+        if(t_process->state() == QProcess::Running)t_process->kill();
+        t_process->waitForFinished(100);
+    }
 }
 
 void DosWrapper::dir_changed(const QString &path)
@@ -149,7 +152,7 @@ void DosWrapper::dir_changed(const QString &path)
         auto file = QFile(t_data.out_acc_name.c_str());
         if(file.exists()){
              step(state::input);
-              return;
+             return;
         }
     }
 }
@@ -167,12 +170,14 @@ void DosWrapper::step(DosWrapper::state s)
             string s;
             getline(cin, s);
             if (empty(s)){
-                if(!cin) emit finished();
                 QCoreApplication::processEvents();
+                if(!cin) {
+                    emit finished();
+                    return;
+                }
                 continue;
             }
             send(s + "\n\n", t_data.out_name, t_data.out_acc_name);
-            if(!cin){t_true_empty =true;}
             break;
         }
         if(!cin || !cout) {
@@ -187,8 +192,21 @@ void DosWrapper::step(DosWrapper::state s)
 void DosWrapper::t_create_conf()
 {
     QFile::remove((t_data.sinchro_path + "/sinchro/conf.conf").c_str());
+    if(!is_exist_dir((t_data.sinchro_path + "/sinchro").c_str())){
+        auto d = QDir(t_data.sinchro_path.c_str());
+        if(!d.mkdir("sinchro")){
+            cout << t_data.sinchro_path << "/sinchro creating error" <<endl;
+            emit finished();
+            return;
+        }
+    }
+    std::this_thread::sleep_for(std::chrono::milliseconds(100));
     QFile::copy((t_data.sinchro_path + "/template.conf").c_str(),(t_data.sinchro_path + "/sinchro/conf.conf").c_str());
     auto conf = ofstream(t_data.sinchro_path + "/sinchro/conf.conf", std::ios::app);
+
+    conf << "cycles=" << t_data.dos_cpu_cycles << endl;
+    conf << endl;
+    conf << "[autoexec]" << endl;
     conf << endl;
     conf << "MOUNT " << t_data.disk_letter << ": " << t_data.disk_path << endl;
     conf << t_data.disk_letter << ":" << endl;
